[MYDA-705][card] Update Entity property type (#377)
* Update Entity property type

* Update point and revolving

* Update transactionResponseCode

* Update test code

* Update revolving and point

+++ card/src/main/java/com/banksalad/collectmydata/card/card/BillDetailResponseHelper.java
@@ -65,7 +65,7 @@ public class BillDetailResponseHelper implements BillTransactionResponseHelper<B
           billDetailEntity.setOrganizationId(executionContext.getOrganizationId());
           billDetailEntity.setChargeMonth(billBasic.getChargeMonth());
           billDetailEntity.setSeqno(billBasic.getSeqno());
-          billDetailEntity.setBillDetailNo(nextBillDetailNo.incrementAndGet());
+          billDetailEntity.setBillDetailNo((short) nextBillDetailNo.incrementAndGet());
           return billDetailEntity;
         }).collect(Collectors.toList());
