+++ card/src/main/java/com/banksalad/collectmydata/card/common/db/entity/BillDetailHistoryEntity.java
@@ -43,7 +43,7 @@ public class BillDetailHistoryEntity extends BaseEntity {
   private String seqno;
 
   @Column(nullable = false)
-  private Integer billDetailNo;
+  private Short billDetailNo;
 
   @Column(nullable = false)
   private String cardId;
